Function aas_benchmark::algorithms::dfa::dfa_with_lps_delta

pub fn dfa_with_lps_delta(
    pattern: &[u8], 
    text: &[u8], 
    delta: fn(_: isize, _: u8, _: &[u8], _: &Vec<isize>) -> isize, 
    lps: &Vec<isize>, 
    i0: usize
) -> Option<usize>

Simulates a deterministic finite automation (DFA) using a transition function delta which takes an lps table (used for KMP algorithm).

The function iterates over each character in the text, starting at i0, and applies the delta function (with given lps table) to each of those characters.

The algorithm terminates when the currently active state q is the last state. It then returns for which character the final state has been reached and therefore and occurrence has been found. If the finals tate could not be reached by iterating over each character in text, None is returned.

Very similar to fn dfa_with_delta.