Module aas_benchmark::match_algorithm

Structs

ALGORITHMS
List of existing algorithms and their internal names

Functions

algorithm_name
Returns the pretty formatted name of an algorithm matching the given name.

match_algorithm
Returns the algorithm function matching the given name.

match_algorithms
Returns the algorithm functions and names matching the given names.

Type Definitions

Algorithm
The signature of a pattern matching algorithm