Module aas_benchmark::measure

Modules

csv_record

measure_result

Functions

calculate_avg_duration
A function to calculate the average duration of a Vec<Duration> in milliseconds.

measure
A function to measure the runtime of an algorithm.

measure_multiple
A function to measure the runtimes of multiple executions of an algorithm.

measure_multiple_different_patterns
Measures the runtimes of multiple executions of an algorithm using a different patterns.